set default of `valid_idx` to 0 in `get_reward`
Giving a value to `valid_idx` would always be overwritten by the if block.
I think the author intended to write:

```python
if not valid_idx:
    valid_idx = 0
```

But forgot the `not` in the condition. In fact instead of adding the `not` we can get rid of the if block and set the `valid_idx` parameter to a default value of 0.

Before this update, omitting to give the optional parameter `valid_idx` would produce an exception.